Fix exports on DeepImageChannel and FlatImageChannel.
If the whole class isn't exported, the typeinfo doesn't get exported,
and so dynamic casting into those classes will not work.

Note that this is a manual port of bb3bb3f from the RB-3.1 branch.
I was unable to do a cherry-pick due to conflicts from minor text
formatting differences between branches.